Factor out tools for filling and rendering email templates
